# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "System Initiative API",
    "description": "The API Server for interacting with a System Initiative workspace",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "This API Server"
    }
  ],
  "paths": {
...
    "/v1/w/{workspace_id}/change-sets/{change_set_id}/schema": {
      "get": {
        "tags": [
          "schemas"
        ],
        "operationId": "list_schemas",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Schemas listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSchemaV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/w/{workspace_id}/change-sets/{change_set_id}/schema/find": {
      "get": {
        "tags": [
          "schemas"
        ],
        "operationId": "find_schema",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "schema",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          },
          {
            "name": "schemaId",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "Schema retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindSchemaV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Schema not found"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/w/{workspace_id}/change-sets/{change_set_id}/schema/{schema_id}": {
      "get": {
        "tags": [
          "schemas"
        ],
        "operationId": "get_schema",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "schema_id",
            "in": "path",
            "description": "Schema identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Schema retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchemaV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Schema not found"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/w/{workspace_id}/change-sets/{change_set_id}/schema/{schema_id}/variant/default": {
      "get": {
        "tags": [
          "schemas"
        ],
        "operationId": "get_default_variant",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "schema_id",
            "in": "path",
            "description": "Schema identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Schema variant retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchemaVariantV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Schema variant not found"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/w/{workspace_id}/change-sets/{change_set_id}/schema/{schema_id}/variant/{schema_variant_id}": {
      "get": {
        "tags": [
          "schemas"
        ],
        "operationId": "get_variant",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "schema_id",
            "in": "path",
            "description": "Schema identifier",
            "required": true
          },
          {
            "name": "schema_variant_id",
            "in": "path",
            "description": "Schema variant identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Schema variant retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSchemaVariantV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Schema variant not found"
          },
          "412": {
            "description": "Schema variant not found for schema"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/w/{workspace_id}/change-sets/{change_set_id}/secrets/{secret_id}": {
      "put": {
        "tags": [
          "secrets"
        ],
        "operationId": "update_secret",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "secret_id",
            "in": "path",
            "description": "Secret identifier",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSecretV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Secret updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSecretV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Secret not found"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "secrets"
        ],
        "operationId": "delete_secret",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace identifier",
            "required": true
          },
          {
            "name": "change_set_id",
            "in": "path",
            "description": "Change set identifier",
            "required": true
          },
          {
            "name": "secret_id",
            "in": "path",
            "description": "Secret identifier",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Secret deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSecretV1Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "404": {
            "description": "Secret not found"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Standard error response format for v1 API",
        "required": [
          "message",
          "statusCode"
        ],
        "properties": {
          "code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "example": 4001
          },
          "message": {
            "type": "string",
            "example": "Invalid request data"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32",
            "example": 422,
            "minimum": 0
          }
        },
        "example": {
          "message": "Invalid request data",
          "statusCode": 422,
          "code": 4001
        }
      },
...
      "DeleteSecretV1Response": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "FindSchemaV1Response": {
        "type": "object",
        "required": [
          "schemaName",
          "schemaId",
          "category",
          "installed"
        ],
        "properties": {
          "category": {
            "type": "string"
          },
          "installed": {
            "type": "boolean"
          },
          "schemaId": {
            "type": "string"
          },
          "schemaName": {
            "type": "string"
          }
        }
      },
...
      "GetSchemaV1Response": {
        "type": "object",
        "required": [
          "name",
          "defaultVariantId",
          "variantIds"
        ],
        "properties": {
          "defaultVariantId": {
            "type": "string",
            "example": "01H9ZQD35JPMBGHH69BT0Q79VZ"
          },
          "name": {
            "type": "string",
            "example": "AWS::EC2::Instance"
          },
          "variantIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "01H9ZQD35JPMBGHH69BT0Q79VZ",
              "01H9ZQD35JPMBGHH69BT0Q79VY"
            ]
          }
        }
      },
      "GetSchemaVariantV1Response": {
        "type": "object",
        "required": [
          "variantId",
          "displayName",
          "category",
          "color",
          "isLocked",
          "description",
          "link",
          "assetFuncId",
          "variantFuncIds",
          "isDefaultVariant"
        ],
        "properties": {
          "assetFuncId": {
            "type": "string",
            "example": "01H9ZQD35JPMBGHH69BT0Q75XY"
          },
          "category": {
            "type": "string",
            "example": "AWS::EC2"
          },
          "color": {
            "type": "string",
            "example": "#FF5733"
          },
          "description": {
            "type": "string",
            "example": "Amazon EC2 Instance resource type"
          },
          "displayName": {
            "type": "string",
            "example": "AWS EC2 Instance"
          },
          "isDefaultVariant": {
            "type": "boolean",
            "example": true
          },
          "isLocked": {
            "type": "boolean",
            "example": false
          },
          "link": {
            "type": "string",
            "example": "https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-ec2-instance.html"
          },
          "variantFuncIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "01H9ZQD35JPMBGHH69BT0Q75AA",
              "01H9ZQD35JPMBGHH69BT0Q75BB"
            ]
          },
          "variantId": {
            "type": "string",
            "example": "01H9ZQD35JPMBGHH69BT0Q79VZ"
          }
        }
      },
...
      "ListSchemaV1Response": {
        "type": "object",
        "required": [
          "schemas"
        ],
        "properties": {
          "schemas": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "example": "[{\"schemaId\":\"01H9ZQD35JPMBGHH69BT0Q79VY\",\"schemaName\":\"AWS::EC2::Instance\",\"category\":\"AWS::EC2\",\"installed\": \"true\"}]"
          }
        }
      },
...
      "SecretV1": {
        "type": "object",
        "required": [
          "id",
          "name",
          "definition",
          "description"
        ],
        "properties": {
          "definition": {
            "type": "string",
            "example": "aws_credentials"
          },
          "description": {
            "type": "string",
            "example": "AWS credentials for production environment"
          },
          "id": {
            "type": "string",
            "example": "01HAXYZF3GC9CYA6ZVSM3E4YHH"
          },
          "name": {
            "type": "string",
            "example": "Production AWS Key"
          }
        }
      },
...
      "UpdateSecretV1Request": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string",
            "example": "Updated AWS Secret Key for EC2 access"
          },
          "id": {
            "type": "string",
            "example": "01HAXYZF3GC9CYA6ZVSM3E4YHH"
          },
          "name": {
            "type": "string",
            "example": "AWS Access Key"
          },
          "rawData": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            },
            "example": {
              "access_key_id": "AKIAIOSFODNN7EXAMPLE",
              "secret_access_key": "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY",
              "region": "us-west-2",
              "default_output": "json"
            }
          }
        }
      },
      "UpdateSecretV1Response": {
        "type": "object",
        "required": [
          "secret"
        ],
        "properties": {
          "secret": {
            "$ref": "#/components/schemas/SecretV1"
          }
        }
      },
...
    }
  },
...
}